trying new approach for ai
