+++ config/openaiclient.js
@@ -1,5 +1,7 @@
-const openai = require('openai');
+const Openai = require('openai');
 
-openai.apiKey = process.env.OPENAI_API_KEY;
+const openai = new OpenAI({
+  apiKey: process.env.OPENAI_API_KEY,
+});
 
 module.exports = openai;
