+++ routes/quiz.js
@@ -4,6 +4,6 @@ const quizController = require("../controllers/quiz");
 
 // Quiz Routes
 router.get("/", quizController.getQuiz);
-router.post("/handleQuizStep", quizController.handleQuizStep); 
+router.post("/api", quizController.aiPrompt);
 
 module.exports = router;
